@@ -26,35 +26,35 @@
     "ngx-color-picker": "^6.3.3",
     "popper.js": "^1.14.3",
     "ts-node": "^6.0.0",
-    "typescript": "^2.6.2",
+    "typescript": "2.7.2",
     "validator": "^9.4.1"
   },
   "peerDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "@types/node": ">= 7.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^3.1.0",
     "bootstrap": "^4.0.0",
-    "modelsafe": "^2.0.0",
-    "rxjs": "^5.0.0",
-    "zone.js": "^0.8.0"
+    "modelsafe": "^2.0.2",
+    "rxjs": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "@types/chai": "^3.4.35",
     "@types/chai-as-promised": "0.0.31",
     "@types/mocha": "^2.2.33",
     "@types/node": ">= 7.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^3.1.0",
     "bootstrap": "^4.0.0",
     "chai": "^3.5.0",
     "chai-as-promised": "^7.0.0",
@@ -62,11 +62,11 @@
     "husky": "^0.13.2",
     "jsverify": "^0.8.2",
     "mocha": "^3.2.0",
-    "modelsafe": "^2.0.0",
-    "rxjs": "^5.0.0",
-    "tslint": "^5.0.0",
+    "modelsafe": "^2.0.2",
+    "rxjs": "^6.2.2",
+    "tslint": "~5.9.1",
     "tslint-config-ccs": "^0.6.2",
-    "zone.js": "^0.8.0"
+    "zone.js": "^0.8.26"
   },
   "scripts": {
     "start": "npm run build",
